+++ sheriff/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for sheriff
 
+## 0.2.1.1
+* Add module name check for self recursive function calls
+* Add top level module name resolving for instance functions
+* Add test cases for calling function with same name but from different module
+* Add State based flow implementation for infinite recursion detection
+
 ## 0.2.1.0
 * Add infinite recursion rule
 * Add test suite for infinite recursion rule detection